Added "Regions"
a "Region" is an group of all territories that belong to 1 player.  Regions can have "isolated" territories and "border" territories.  When it comes time for troop movement... the AI moves all troops (except for 1) from the isolated territories and evenly (as close to evenly as numerically possible) distributes the troops to border territories.